+++ packages/nesd/lib/nes/cartridge/mapper/mapper_state.dart
@@ -15,6 +15,7 @@ import 'package:nesd/nes/cartridge/mapper/namco108_state.dart';
 import 'package:nesd/nes/cartridge/mapper/namco163_state.dart';
 import 'package:nesd/nes/cartridge/mapper/nrom_state.dart';
 import 'package:nesd/nes/cartridge/mapper/single_prg_bank_state.dart';
+import 'package:nesd/nes/cartridge/mapper/unrom512_state.dart';
 
 abstract class MapperState {
   const MapperState({required this.id});
@@ -42,6 +43,7 @@ abstract class MapperState {
       9 => MMC2State.deserialize(reader),
       16 => BandaiFCGState.deserialize(reader),
       19 => Namco163State.deserialize(reader),
+      30 => UNROM512State.deserialize(reader),
       45 => Mapper45State.deserialize(reader),
       66 => GxROMState.deserialize(reader),
       71 => SinglePrgBankState.deserialize(reader, 71),